% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{anomaly_detection_performance_split}
\alias{anomaly_detection_performance_split}
\title{Anomaly detection performance calculator}
\usage{
anomaly_detection_performance_split(
  estimated_anomalies,
  true_anomalies_point,
  true_anomalies_contextual
)
}
\arguments{
\item{estimated_anomalies}{the anomalies detected by procedure}

\item{true_anomalies_point}{the true point anomalies locations}

\item{true_anomalies_contextual}{the true contextual anomaly locations}
}
\value{
a list containing the performance of detection procedure overall and split into point and contextual anomalies
}
\description{
This function measures the performance of the anomaly detection procedure in terms of
F1, Recall and precision. Additionally it splits this up into correctly identified
point and contextual anomalies.
}
\examples{
anomaly_detection_performance_split(c(1,5,4),c(1),c(5))

}
