% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{conformal_prediction_p_val}
\alias{conformal_prediction_p_val}
\title{Calculate Conformal P-Values}
\usage{
conformal_prediction_p_val(non_conform_measures)
}
\arguments{
\item{non_conform_measures}{sequence of non conformity measures}
}
\value{
a p-value (between 0 and 1) for the final value in the input sequence
}
\description{
This function takes a sequence of non conformity measures (in this case,
the absolute forecasting error is used) and produces a p-value for the
final value in the sequence.
}
