% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{data_simulation_function}
\alias{data_simulation_function}
\title{Simulate periodic data with point and contextual anomalies}
\usage{
data_simulation_function(
  n,
  burnin_period,
  no_additive_long,
  no_subtractive_long,
  no_point,
  day_freq = 1440,
  ar_coef = 0.05,
  ma_coef = 0.05
)
}
\arguments{
\item{n}{the length of the simulated sequence}

\item{burnin_period}{length of initial burn-in without anomalies}

\item{no_additive_long}{number of additive contextual anomalies}

\item{no_subtractive_long}{number of subtractive contextual anomalies}

\item{no_point}{number of point anomalies}

\item{day_freq}{the period length of the periodic cycle which is automatically set to 1440}

\item{ar_coef}{the AR parameter coefficient default = 0.05}

\item{ma_coef}{the MA parameter coefficient default = 0.05}
}
\value{
a list containing the series with and without anomalies added, and locations of the point and contextual anomalies
}
\description{
This function produces a periodic time series containing the specified number of
point and contextual anomalies. The data uses a sine curve for the long-term
component and an ARMA(1,1) process for the short-term structure. Gaussian noise
is added to the series
}
\examples{
data_simulation_function(1440,100,2,2,30)

}
