% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{fda_mean_weighted_extreme}
\alias{fda_mean_weighted_extreme}
\title{FDA model calculator for data with anomalies}
\usage{
fda_mean_weighted_extreme(
  datamat,
  rangex,
  nb,
  no,
  obsweights,
  extreme_days,
  extreme_bins,
  includebpoints = FALSE,
  bpoints = NULL
)
}
\arguments{
\item{datamat}{a n*p matrix containing all previous periods of the series where n is
the length of the period and p is the total number of periods observed}

\item{rangex}{vector c(1:n), the point numbers in the period}

\item{nb}{number of basis functions to use}

\item{no}{order of basis functions}

\item{obsweights}{vector of weights for each column of the data}

\item{extreme_days}{a vector of values corresponds to the column of anomalous points
to be excluded from the model}

\item{extreme_bins}{a vector of values corresponding to the rows of anomalous points
to be excluded from the model}

\item{includebpoints}{A booliean. If TRUE then must specify breakpoints otherwise do not}

\item{bpoints}{Optional vector (required if includebpoints is TRUE). This details the breakpoints
to use for the basis functions}
}
\value{
List containing average, sd and upper and lower ranges.
}
\description{
This function produces a weighted FDA curve for the input series but excludes
anomalous points from its calculation.
The coefficients of the b-spline are calculated using poisson regression.
This gives the average curve, sd and a prediction interval for the curve.
}
