% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{fda_mean_weighted_extreme_update}
\alias{fda_mean_weighted_extreme_update}
\title{FDA model calculator for data with anomalies for a single update from previous (sequential).}
\usage{
fda_mean_weighted_extreme_update(
  current_data,
  current_lambda,
  sum_weights_vec,
  forecast_values,
  rangex,
  nb,
  no,
  which_anomalous,
  includebpoints = FALSE,
  bpoints = NULL
)
}
\arguments{
\item{current_data}{a vector detailing the data from the latest fully observed cycle}

\item{current_lambda}{the latest lambda estimate}

\item{sum_weights_vec}{the weight for each time within the cycle (these weights are
different for each point as anomalous points are excluded hence have weight 0)}

\item{forecast_values}{these are the corresponding forecast values for input current data}

\item{rangex}{vector c(1:n), the point numbers in the period}

\item{nb}{number of basis functions to use}

\item{no}{order of basis functions}

\item{which_anomalous}{details which of the current_data points are considered anomalous (boolean
vector with 1 if anomalous 0 if not)}

\item{includebpoints}{A booliean. If TRUE then must specify breakpoints otherwise do not}

\item{bpoints}{Optional vector (required if includebpoints is TRUE). This details the breakpoints
to use for the basis functions}
}
\value{
List containing average fda curve, the weighted average and the new updated sum_weights_vec which incorporates the
anomalous values
}
\description{
This function produces a weighted FDA curve for the input series but excludes
anomalous points from its calculation.
The coefficients of the b-spline are calculated using linear regression.
This gives the average curve and the new sum of weights for each time within a cycle.
}
