% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDARIMA.R
\name{fda_mean_weighted_poisson}
\alias{fda_mean_weighted_poisson}
\title{FDA model calculator for data without anomalies}
\usage{
fda_mean_weighted_poisson(
  datamat,
  rangex,
  nb,
  no,
  obsweights,
  includebpoints = FALSE,
  bpoints = NULL,
  p = 0.999
)
}
\arguments{
\item{datamat}{a n*p matrix containing all previous periods of the series where n is
the length of the period and p is the total number of periods observed}

\item{rangex}{vector c(1:n), the point numbers in the period}

\item{nb}{number of basis functions to use}

\item{no}{order of basis functions}

\item{obsweights}{vector of weights for each column of the data}

\item{includebpoints}{A booliean. If TRUE then must specify breakpoints otherwise do not}

\item{bpoints}{Optional vector (required if includebpoints is TRUE). This details the breakpoints
to use for the basis functions}

\item{p}{the significance level to use for the prediction interval}
}
\value{
List containing average, sd and upper and lower ranges.
}
\description{
This function produces a weighted FDA curve for the input series.
The coefficients of the b-spline are calculated using poisson regression.
This gives the average curve, sd and a prediction interval for the curve.
}
