% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDARIMA.R
\name{FDARIMA}
\alias{FDARIMA}
\title{FDARIMA anomaly detector}
\usage{
FDARIMA(
  datamat,
  no_days_train,
  max_no_days,
  fda_forgetting_factor = 0.86,
  arima_window_size = 60,
  learning_rate = 0.001,
  threshold_val = 0.005
)
}
\arguments{
\item{datamat}{a n*p matrix containing the time series where n is
the length of the period and p is the total number of periods observed}

\item{no_days_train}{number of periods of data to use as a burnin period}

\item{max_no_days}{this is maximum numbers of days that need to be forecasted using FDA model usually p-1}

\item{fda_forgetting_factor}{this is the intial forgetting factor to be used within exponential weighting (this is updated within the model)}

\item{arima_window_size}{number of historic data points to use to build ARIMA model}

\item{learning_rate}{the learning rate within stochastic gradient descent to update lambda
to be excluded from the model}

\item{threshold_val}{the threshold value used to detect anomalies}
}
\value{
returns list containing a matrix which contains the forecasts, and anomalies detected by each procedure and combined anomalies. Also
outputs a separate vector containing the combined anomalies only
}
\description{
Detects anomalies by separately detecting anomalies using long-term and short-term
models and combining their p-values using Fishers product test statistic to give
final anomaly score that indicates both point and contextual anomalies.
}
