% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDARIMA.R
\name{FDARIMA_initial}
\alias{FDARIMA_initial}
\title{FDARIMA anomaly detector initialiser}
\usage{
FDARIMA_initial(
  datamat,
  frequency_of_data,
  no_cycles_train,
  fda_forgetting_factor = 0.99,
  arima_window_size = 60,
  learning_rate = 0.001,
  threshold_val = 0.005,
  sliding_window_size = 14400
)
}
\arguments{
\item{datamat}{a n*p matrix containing the time series where n is
the length of the period and p is the total number of periods observed}

\item{frequency_of_data}{this is the length of the period nature of the data}

\item{no_cycles_train}{number of periods of data to use as a burnin period}

\item{fda_forgetting_factor}{this is the intial forgetting factor to be used within exponential weighting (this is updated within the model)}

\item{arima_window_size}{number of historic data points to use to build ARIMA model}

\item{learning_rate}{the learning rate within stochastic gradient descent to update lambda
to be excluded from the model}

\item{threshold_val}{the threshold value used to detect anomalies}

\item{sliding_window_size}{the size of the sliding window for conformal prediction. This will determine whether can}
}
\value{
returns list containing a matrix which contains the forecasts, and anomalies detected by each procedure and combined anomalies. Also
outputs a separate vector containing the combined anomalies only
}
\description{
Detects anomalies by separately detecting anomalies using long-term and short-term
models and combining their p-values using Fishers product test statistic to give
final anomaly score that indicates both point and contextual anomalies. This is the funciton to initiate
running in real time. Needed to initially start the model. Anomalies will not be detected until the conformal
prediction sliding window is filled (but model forecasts will still be computed).
}
