% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDARIMA.R
\name{newestAff}
\alias{newestAff}
\title{Uses Shermann-Morrison theorem to update A matrix}
\usage{
newestAff(A, v, lambda)
}
\arguments{
\item{A}{A matrix}

\item{v}{new values}

\item{lambda}{forgetting factor}
}
\value{
new A matrix
}
\description{
Updates A matrix for sequential updates of weighted linear regression
}
