% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDARIMA.R
\name{newestbetaff}
\alias{newestbetaff}
\title{Uses Shermann-Morrison theorem to update paramter vector}
\usage{
newestbetaff(beta, A, v, y, lambda)
}
\arguments{
\item{beta}{parameter vector}

\item{A}{A matrix}

\item{v}{new values}

\item{y}{new response}

\item{lambda}{forgetting factor}
}
\value{
new parameter vector beta
}
\description{
Updates parameter vector matrix sequential updates of weighted linear regression
}
