% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDARIMA.R
\name{regression_combination_forecast}
\alias{regression_combination_forecast}
\title{Forecast combiner using weighted linear regression}
\usage{
regression_combination_forecast(
  True_values,
  FDA_values,
  ARIMA_values,
  start_val,
  end_val,
  lambda = 0.98
)
}
\arguments{
\item{True_values}{the real time series}

\item{FDA_values}{the forecasted FDA values for the input series}

\item{ARIMA_values}{the forecasted ARIMA values for the input series}

\item{start_val}{the number of points to initialise with}

\item{end_val}{when to stop the estimation}

\item{lambda}{the fixed forgetting factor to use}
}
\value{
combined forecast values
}
\description{
Combines short-term and long term forecasts using fixed exponentially weighted linear regression.
}
