% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLDecomposition.R
\name{SLD_updater}
\alias{SLD_updater}
\title{SL Decomposition anomaly detector updater}
\usage{
SLD_updater(
  i,
  new_data_point,
  new_data,
  current_data,
  SLD_fda_anomaly,
  previous_FDA_wm,
  current_FDA_wm,
  ARIMA_fcast,
  sum_weights_vec,
  sum_weights_vec_new,
  SLD_fda_error_anomaly,
  SLD_combined_prediction_error_anomaly,
  strange_data_points,
  frequency_of_data,
  fda_forgetting_factor = 0.99,
  arima_window_size = 60,
  learning_rate = 0.001,
  threshold_val = 0.005,
  sliding_window_size = 14400
)
}
\arguments{
\item{i}{the current data point number}

\item{new_data_point}{the next value in the series that is used to update the models and asssess if anomalous}

\item{current_data}{this is the value at the previous time point in the series}

\item{SLD_fda_anomaly}{this is a vector detailing the current FDA forecasts for the current cycle}

\item{previous_FDA_wm}{the weighted average from the previous update (at $i-2$)}

\item{current_FDA_wm}{the weighted average from update at time $i$}

\item{ARIMA_fcast}{this is the ARIMA forecast for time $i$}

\item{sum_weights_vec}{this is a vector detailing the weights for each time within the cycle(for the FDA model) for time i-2. This accounts for anomalous data.}

\item{sum_weights_vec_new}{this is a vector detailing the weights for each time within the cycle(for the FDA model) for time i-1. This accounts for anomalous data.}

\item{SLD_fda_error_anomaly}{the absolute error for the FDA mdoel. This is collected to run the conformal predictor and is kept at the lenght of the sliding window.}

\item{SLD_combined_prediction_error_anomaly}{the absolute error for the combined forecast for SLD. This is collected to run the conformal predictor and is kept at the lenght of the sliding window.}

\item{strange_data_points}{this outlines anomalous points and is a vector of times of anomalous points identified (only require  points in the past cycle).}

\item{frequency_of_data}{this is the length of the period nature of the data}

\item{fda_forgetting_factor}{this is the intial forgetting factor to be used within exponential weighting (this is updated within the model)}

\item{arima_window_size}{number of historic data points to use to build ARIMA model}

\item{learning_rate}{the learning rate within stochastic gradient descent to update lambda
to be excluded from the model}

\item{threshold_val}{the threshold value used to detect anomalies}

\item{sliding_window_size}{the size of the sliding window for conformal prediction. This will determine whether can}
}
\value{
returns list containing a matrix which contains the forecasts, and anomalies detected by each procedure and combined anomalies. Also
outputs a separate vector containing the combined anomalies only
}
\description{
Detects anomalies by separately detecting anomalies using long-term and combined forecast
from decomposition and combining their p-values using Fishers product test statistic to give
final anomaly score that indicates both point and contextual anomalies. This is the funciton to update
existing calculations as new data becomes available.
}
