% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLDecomposition.R
\name{SLDecomposition}
\alias{SLDecomposition}
\title{SL Decomposition anomaly detector}
\usage{
SLDecomposition(
  datamat,
  arima_window = 60,
  no_days_train = 10,
  starting_no_obs = 5,
  fda_forgetting_factor = 0.99,
  learning_rate = 0.001,
  threshold_val = 0.005
)
}
\arguments{
\item{datamat}{a n*p matrix containing the time series where n is
the length of the period and p is the total number of periods observed}

\item{no_days_train}{number of periods of data to use as a burnin period}

\item{starting_no_obs}{this is the number of days to start collecting residuals in long-term process to
forecast the short-term process using ARIMA}

\item{fda_forgetting_factor}{this is the intial forgetting factor to be used within exponential weighting (this is updated within the model)}

\item{learning_rate}{the learning rate within stochastic gradient descent to update lambda
to be excluded from the model}

\item{threshold_val}{the threshold value used to detect anomalies}

\item{arima_window_size}{number of historic data points to use to build ARIMA model}
}
\value{
returns list containing a matrix which contains the forecasts, and anomalies detected by each procedure and combined anomalies. Also
outputs a separate vector containing the combined anomalies only
}
\description{
Detects anomalies by separately detecting anomalies using long-term and combined forecast
from decomposition and combining their p-values using Fishers product test statistic to give
final anomaly score that indicates both point and contextual anomalies.
}
