% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{update_fda_ff_adaptive}
\alias{update_fda_ff_adaptive}
\title{Stochastic gradient descent updater for forgetting factor using adaptive approach}
\usage{
update_fda_ff_adaptive(
  current_lambda,
  previous_weights,
  alpha,
  datamat,
  data_new
)
}
\arguments{
\item{current_lambda}{the current value for lambda}

\item{previous_weights}{the previous weights for the data}

\item{alpha}{the learning rate which controls the change at each update (recommend 10^-3)}

\item{datamat}{a n*p matrix containing all previous periods of the series n is
the length of the period and p is the total number of periods observed}

\item{data_new}{a vector containing the values of the newly observed period}
}
\value{
new forgetting factor
}
\description{
This function updates the forgetting factor using stochastic
gradient descent.
}
