% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{update_fda_ff_adaptive_sequential}
\alias{update_fda_ff_adaptive_sequential}
\title{Stochastic gradient descent updater for forgetting factor using adaptive approach}
\usage{
update_fda_ff_adaptive_sequential(
  new_data,
  current_data,
  current_lambda,
  sum_weights_vec,
  forecast_values_new,
  forecast_values_old,
  alpha,
  which_anomalous
)
}
\arguments{
\item{new_data}{a vector detailing the data from the latest fully observed cycle}

\item{current_data}{a vector detailing the data from the previous fully observed cycle}

\item{current_lambda}{the latest lambda estimate}

\item{sum_weights_vec}{the weight for each time within the cycle (these weights are
different for each point as anomalous points are excluded hence have weight 0). This is from the
previous cycle at this time.}

\item{forecast_values_new}{these are the corresponding forecast values for new_data}

\item{forecast_values_old}{these are the corresponding forecast values for current_data}

\item{alpha}{the learning rate which controls the change at each update (recommend 10^-3)}

\item{which_anomalous}{details which of the current_data points are considered anomalous (boolean
vector with 1 if anomalous 0 if not)}
}
\value{
new forgetting factor
}
\description{
This function updates the forgetting factor using stochastic
gradient descent.
}
